LayoutTests/storage/indexeddb/modern/basic-add-expected.txt

This test does basic testing of IDBObjectStore.add(), making sure that an attempt to overwrite an already-existing key fails with the appropriate error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: Upgrade needed: Old version - 0 New version - 1
ALERT: [object IDBTransaction] - versionchange
ALERT: [object IDBDatabase]
ALERT: put 1 succeeded - key was 'foo'
ALERT: put 2 failed - error
ALERT: get succeeded - key was 'bar'
ALERT: version change transaction completed
PASS successfullyParsed is true

TEST COMPLETE
